Python Project 6(Chrome Dinosour automation)/requirements.txt
MouseInfo==0.1.3
numpy==1.19.1
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
